[nasa/nos3#464] Reorganized and reconfigured EPS checkout to match current checkout standards. May still need some updates
