+++ .gitignore
@@ -1,1 +1,1 @@
-support/build
+fsw/standalone/build
