+++ fsw/cfs/CMakeLists.txt
@@ -8,6 +8,7 @@ include_directories(mission_inc)
 include_directories(platform_inc)
 include_directories(src)
 
+include_directories(../shared)
 include_directories(${hwlib_MISSION_DIR}/fsw/public_inc)
 
 aux_source_directory(src APP_SRC_FILES)
